wsi_preprocessing/requirements.txt
dino
h5py
imageio
imgaug
matplotlib
numpy
opencv_python
openslide_python
pandas
Pillow
pyvips
scikit_image
scikit_learn==1.1.2
scipy==1.7.3
skimage==0.0
tifffile==2020.2.16
timm==0.4.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
zarr==2.3.2
